+++ internal/resolvers/default/resolver.go
@@ -31,12 +31,6 @@ const (
 <span>{{.Description}}</span><hr>
 {{end}}
 <b>URL:</b> {{.URL}}</div>`
-
-	// MaxTitleLength describes the maximum length of a title when returned from the default link resolver
-	MaxTitleLength = 60
-
-	// MaxDescriptionLength describes the maximum length of an og:description when returned from the default link resolver
-	MaxDescriptionLength = 200
 )
 
 var (