+++ internal/resolvers/frankerfacez/resolver.go
@@ -2,8 +2,8 @@ package frankerfacez
 
 import (
 	"errors"
+	"html/template"
 	"regexp"
-	"text/template"
 	"time"
 
 	"github.com/Chatterino/api/pkg/cache"
@@ -11,8 +11,6 @@ import (
 )
 
 const (
-	emoteAPIURL = "https://api.frankerfacez.com/v1/emote/%s"
-
 	thumbnailFormat = "https://cdn.frankerfacez.com/emoticon/%s/4"
 
 	tooltipTemplate = `<div style="text-align: left;">
@@ -22,6 +20,8 @@ const (
 )
 
 var (
+	emoteAPIURL = "https://api.frankerfacez.com/v1/emote/%s"
+
 	// FrankerFaceZ hosts we're doing our smart things on
 	domains = map[string]struct{}{
 		"frankerfacez.com":     {},