+++ internal/resolvers/twitch/load.go
@@ -24,14 +24,14 @@ func load(clipSlug string, r *http.Request) (interface{}, time.Duration, error) 
 
 	var clipHelix = response.Data.Clips[0]
 
-	var createdData, _ = time.Parse("2006-01-02T15:04:05Z", clipHelix.CreatedAt)
+	//var createdData, _ = time.Parse("2006-01-02T15:04:05Z", clipHelix.CreatedAt)
 
 	data := twitchClipsTooltipData{
 		Title:       clipHelix.Title,
 		AuthorName:  clipHelix.CreatorName,
 		ChannelName: clipHelix.BroadcasterName,
-		// Duration: // fmt.Sprintf("%g%s", clip.Duration, "s")
-		CreationDate: createdData.Format("02 Jan 2006"),
+		//Duration: // humanize.DurationSeconds(time.Duration(clip.Duration) * time.Second)
+		CreationDate: humanize.CreationDateRFC3339(clipHelix.CreatedAt),
 		Views:        humanize.Number(uint64(clipHelix.ViewCount)),
 	}