+++ internal/resolvers/twitch/token_refresh.go
@@ -1,0 +1,34 @@
+package twitch
+
+import (
+	"log"
+	"time"
+
+	"github.com/nicklaw5/helix"
+)
+
+func requestAppAccessToken(helixAPI *helix.Client) {
+	// Request app access token
+	response, err := helixAPI.RequestAppAccessToken([]string{})
+
+	if err != nil {
+		log.Fatalf("[Helix] Error requesting app access token: %s , \n %s", err.Error(), response.Error)
+	}
+
+	log.Printf("[Helix] Requested access token, status: %d, expires in: %d", response.StatusCode, response.Data.ExpiresIn)
+	helixAPI.SetAppAccessToken(response.Data.AccessToken)
+
+	// Refresh app access token every 24 hours
+	ticker := time.NewTicker(24 * time.Hour)
+
+	for range ticker.C {
+		response, err := helixAPI.RequestAppAccessToken([]string{})
+		if err != nil {
+			log.Printf("[Helix] Failed to refresh app access token, status: %d", response.StatusCode)
+			continue
+		}
+		log.Printf("[Helix] Requested access token from ticker, status: %d, expires in: %d", response.StatusCode, response.Data.ExpiresIn)
+
+		helixAPI.SetAppAccessToken(response.Data.AccessToken)
+	}
+}