+++ internal/resolvers/wikipedia/resolver.go
@@ -2,8 +2,8 @@ package wikipedia
 
 import (
 	"errors"
+	"html/template"
 	"regexp"
-	"text/template"
 	"time"
 
 	"github.com/Chatterino/api/pkg/cache"
@@ -20,15 +20,8 @@ var (
 
 	errLocaleMatch = errors.New("could not find locale from URL")
 	errTitleMatch  = errors.New("could not find title from URL")
-)
 
-const (
 	endpointURL = "https://%s.wikipedia.org/api/rest_v1/page/summary/%s?redirect=false"
-
-	// TODO: Replace these with shared constants once implemented
-	maxTitleLength       = 60
-	maxDescriptionLength = 60
-	maxExtractLength     = 200
 )
 
 func New() (resolvers []resolver.CustomURLManager) {